init: start of task implementation
